Upcoming Lodha Bangalore Offers Affordable Homes
Lodha Group is one of the most renowned builders in Bengaluru. This residential project offers Fully AC 2 & 3 BHK Premium Residences with Landscape design based on the concept of 'LIVING IN AN ORCHARD'. 

It has Over 1.5 lac sq. ft of outdoor spaces & Full height windows to maximize natural light and ventilation. There is an entrance gateway & security room, a central plaza with a landscape, and a kids’ play range. The homes are the perfect combination of contemporary architecture and features to provide a comfortable living life in Bangalore City.

Bengaluru is also called the Silicon Valley of India and has seen immense demand over the last few years due to the establishment of the IT industry as well as start-ups in the State. A lot of people have migrated to the city in search of job chances. The demand for residential projects in Bangalore has seen immense growth. There is an intercom to the apartment from the security cabin through the service earner.

The complex has covered car parking, on-call maintenance, visitors lobby, health club with multipurpose exercise room, clubhouse with multipurpose hall and indoor games, kids’ play area, landscaped garden, 24 hrs. power backup for the common ranges. The property has round-the-clock security, a waste management system, etc. It is one of the best projects by Lodha Group. Strategically located the property is near all leading hospitals, schools, colleges, banks, ATMs, & so on. This makes it one of the most desirable properties for home purchasers. So, book your home today.

For More Details:
Visit Here: https://www.lodhadeveloper.co.in/residential/bangalore/lodha-bangalore/